+++ Check.ID.Pi.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-cd /home/pi/checkid
+cd /home/pi/SensorFusion/
 
-source checkid-env/bin/activate
+source SF-env/bin/activate
 
-python3 TFLite_detection_webcam.py --modeldir=Sample_TFLite_model
+cd /home/pi/SensorFusion/checkid
+
+python3 TFLite_detection_webcam_old.py --modeldir=Sample_TFLite_model