+++ Thermal01.C.sh
@@ -1,8 +1,9 @@
-# Pushkar Khairnar
-#/bash_file/tflite_thermal_live_with_coral_run
+#!/bin/bash
 
-cd /home/pi/thermal01/
+cd /home/pi/SensorFusion/
 
-source thermal01-env/bin/activate
+source SF-env/bin/activate
+
+cd /home/pi/SensorFusion/thermal01/
 
 python3 simpleVideoCamera.py --modeldir=Sample_TFLite_model --edgetpu