+++ Thermal01.seekware.sh
@@ -1,7 +1,6 @@
-# Pushkar Khairnar
-#/bash_file/tflite_thermal_live_with_coral_run
+#!/bin/bash
 
-cd /home/pi/thermal01/seek_to_csv/seekware-server/build/
+cd /home/pi/SensorFusion/thermal01/seek_to_csv/seekware-server/build/
 
 #source thermal01-env/bin/activate