+++ launch.sh
@@ -1,10 +1,9 @@
 #!/usr/bin/python3
 
-python3 -mwebbrowser http://localhost:5000
+cd /home/pi/SensorFusion/
 
-cd /home/pi/Demo90
+source SF-env/bin/activate
 
-source Demo90-env/bin/activate
+cd /home/pi/SensorFusion/Demo90
 
-python3 TFLite_detection_webcam_api.py --modeldir=Sample_TFLite_model --edgetpu
-
+python3 TFLite_detection_webcam_api.py --modeldir=Sample_TFLite_model --edgetpu |& python3 -mwebbrowser http://localhost:5000